backend/foodgram/requirements.txt
asgiref==3.5.0
Django==3.2.11
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.10.1
Markdown==3.3.6
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
zipp==3.7.0
